Refactor: 인코딩 디코딩 함수 utils로 뽑기
