+++ src/utils/bytesToBase64.ts
@@ -1,0 +1,10 @@
+// 인코딩
+// 예시
+// input: new Uint8Array([72, 101, 108, 108, 111])
+// output: SGVsbG8=  // "Hello"
+
+const bytesToBase64 = (bytes: Uint8Array): string => {
+  const binString = String.fromCodePoint(...Array.from(bytes));
+  return btoa(binString);
+};
+export default bytesToBase64;
