+++ src/utils/index.ts
@@ -3,5 +3,7 @@ import { isEmailCheck } from "./emailCheck";
 import getTxtColorByBgColor from "./getTxtColorByBgColor";
 import isNotUser from "./isNotUser";
 import convertUrlToFile from "./convertUrlToFile";
+import base64ToBytes from "./base64ToBytes";
+import bytesToBase64 from "./bytesToBase64";
 
-export { isPasswordCheck, isPasswordDoubleCheck, isEmailCheck, getTxtColorByBgColor, isNotUser, convertUrlToFile };
+export { isPasswordCheck, isPasswordDoubleCheck, isEmailCheck, getTxtColorByBgColor, isNotUser, convertUrlToFile, base64ToBytes, bytesToBase64 };
